requirement.txt
telethon==1.19.5
python-binance==0.7.9
pandas==1.2.2